Changes lang on app mount
