+++ app/javascript/dashboard/helper/commons.js
@@ -9,3 +9,13 @@ export default () => {
     });
   }
 };
+
+export const accountIdFromUrl = () => {
+  const isInsideAccountScopedURLs = window.location.pathname.includes(
+    '/app/accounts'
+  );
+  const accountId = isInsideAccountScopedURLs
+    ? window.location.pathname.split('/')[3]
+    : '';
+  return accountId;
+};
